Fix mod not working on Linux/Proton
For some reason, coroutines created in
OnMagicNumbersAndWorldSeedInitialized can not be resumed in
OnWorldPreUpdate. At least, that's what I observed on my machine.

This means that the loop checking for the connection doesn't run which
stops the mod from initialising properly. I suspect this doesn't impact
Windows because it instantly knows whether the connection is good or bad
so it never enters the loop.

The rest of the code couldn't quite cope with the initialisation being
delayed when it needs to enter that loop. I added a `nil` check on the
`is_server` variable to fix that.